using console.table to display results.

+++ seeds.sql
@@ -8,7 +8,7 @@ CREATE TABLE products(
     department_name VARCHAR(50) NOT NULL,
     price DECIMAL(10,2) NOT NULL,
     stock_quantity INTEGER NOT NULL,
-    product_sales INTEGER NULL DEFAULT 0,
+    product_sales DECIMAL(10, 2) NULL DEFAULT 0,
     PRIMARY KEY(item_id)
 );
